Better set up results for location api
It's somewhat awkward to get a huge listing of trains and need
to find their routes individually. This is fundamentally a problem
because of how HTTParty (and in turn, MultiXml) parses the semi-invalid
XML we get from the CTA TrainTracker API.

To alleviate, parse out the route names and return a hash with the route
names as the top-most key.