Add CustomerAlerts API Support
Add support to query the CustomerAlerts API. This API has two calls, one
for querying the status of a route and one for querying alerts along the
system.

Add the ability to retrieve a list of all routes from a
cta_routes.csv, and filter them to either train or bus routes if
desired. While the BusTracker API allows a list of routes to be
retrieved dynamically, it doesn't include train routes - and it is
possible that someone might want this data without signing up for a
BusTracker key (as a key is not required for CustomerAlerts).

Take the existing stop_table, stops, and stations methods and extract
them to a shared class that can be used by both the TrainTracker and
CustomerAlerts API classes. I also fixed a minor bug that was introduced
since the stations method was refering an outdated name in the latest
version of the cta_L_stops.csv file.